orders updated

+++ database/migrations/2019_12_29_102925_create_monkcommerce_orders_status_table.php
@@ -13,10 +13,10 @@ class CreateMonkcommerceOrdersStatusTable extends Migration
      */
     public function up()
     {
-        Schema::create('mc_orders_status', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->string('status');
-        });
+      Schema::create('mc_orders_status', function (Blueprint $table) {
+          $table->increments('id');
+          $table->string('status');
+      });
     }
 
     /**
